Support for wildcards + some refactoring.
