+++ eldar/regex.py
@@ -1,0 +1,2 @@
+WORD_REGEX = r'([\w]+|[,?;.:\/!()\[\]\'"’\\><+-=])'
+WILD_CARD_REGEX = r'\w*'
